+++ src/devices/mod.rs
@@ -1,6 +1,5 @@
 pub mod pcf8574;
 pub mod pcf8575;
-mod set_pin;
 
-#[cfg(feature = "unproven")]
 mod get_pin;
+mod set_pin;