+++ src/devices/pcf8575.rs
@@ -1,5 +1,5 @@
 use core::cell;
-use embedded_hal::blocking::i2c::{Read, Write};
+use embedded_hal::i2c::I2c;
 
 use crate::split_pins::pcf8575;
 use crate::{Error, PinFlag, SlaveAddr};
@@ -23,7 +23,7 @@ pub(crate) struct Pcf8575Data<I2C> {
 
 impl<I2C, E> Pcf8575<I2C>
 where
-    I2C: Write<Error = E>,
+    I2C: I2c<Error = E>,
 {
     /// Create new instance of the PCF8575 device
     pub fn new(i2c: I2C, address: SlaveAddr) -> Self {
@@ -94,7 +94,7 @@ where
 
 impl<I2C, E> Pcf8575<I2C>
 where
-    I2C: Read<Error = E> + Write<Error = E>,
+    I2C: I2c<Error = E>,
 {
     /// Get the status of the selected I/O pins.
     /// The mask of the pins to be read can be created with a combination of