+++ src/devices/set_pin.rs
@@ -1,6 +1,6 @@
 use super::super::split_pins;
 use super::super::{Error, Pcf8574, Pcf8574a, Pcf8575, PinFlag};
-use embedded_hal::blocking::i2c::Write;
+use embedded_hal::i2c::I2c;
 
 macro_rules! pcf8574_set_pin_impl {
     ( $( $device_name:ident ),+ ) => {
@@ -13,7 +13,8 @@ macro_rules! pcf8574_set_pin_impl {
             // Again, this is only internal so users cannot misuse it.
             impl<I2C, E> split_pins::SetPin<E> for $device_name<I2C>
             where
-                I2C: Write<Error = E>
+                I2C: I2c<Error = E>,
+                E: core::fmt::Debug
             {
                 fn set_pin_high(&self, pin_flag: PinFlag) -> Result<(), Error<E>> {
                     self.do_on_acquired(|dev|{
@@ -37,7 +38,8 @@ pcf8574_set_pin_impl!(Pcf8574, Pcf8574a);
 
 impl<I2C, E> split_pins::SetPin<E> for Pcf8575<I2C>
 where
-    I2C: Write<Error = E>,
+    I2C: I2c<Error = E>,
+    E: core::fmt::Debug,
 {
     fn set_pin_high(&self, pin_flag: PinFlag) -> Result<(), Error<E>> {
         self.do_on_acquired(|dev| {