+++ tests/pcf8574.rs
@@ -1,4 +1,4 @@
-use embedded_hal_mock::i2c::{Mock as I2cMock, Transaction as I2cTrans};
+use embedded_hal_mock::eh1::i2c::{Mock as I2cMock, Transaction as I2cTrans};
 use pcf857x::{Error, Pcf8574, Pcf8574a, PinFlag, SlaveAddr};
 mod base;
 
@@ -136,7 +136,7 @@ macro_rules! pcf8574_pin_test {
     ($px:ident, $value:expr, $default_address:expr) => {
         mod $px {
             use super::*;
-            #[cfg(feature = "unproven")]
+
             use pcf857x::InputPin;
             use pcf857x::OutputPin;
 
@@ -166,7 +166,6 @@ macro_rules! pcf8574_pin_test {
                 expander.destroy().done();
             }
 
-            #[cfg(feature = "unproven")]
             #[test]
             fn can_split_and_get_is_high() {
                 let transactions = [
@@ -175,13 +174,12 @@ macro_rules! pcf8574_pin_test {
                 ];
                 let expander = new(&transactions);
                 {
-                    let parts = expander.split();
+                    let mut parts = expander.split();
                     assert!(parts.$px.is_high().unwrap());
                 }
                 expander.destroy().done();
             }
 
-            #[cfg(feature = "unproven")]
             #[test]
             fn can_split_and_get_is_low() {
                 let transactions = [
@@ -190,7 +188,7 @@ macro_rules! pcf8574_pin_test {
                 ];
                 let expander = new(&transactions);
                 {
-                    let parts = expander.split();
+                    let mut parts = expander.split();
                     assert!(parts.$px.is_low().unwrap());
                 }
                 expander.destroy().done();